requirements.txt
Pillow
transformers
qwen_vl_utils
torch
torchvision